5.3 Parcial

+++ Trabajo Integrador/src/GUI/Camiones/PantallaCrearCamion.java
@@ -146,7 +146,7 @@ public class PantallaCrearCamion implements ActionListener {
             if (!marca.getText().isEmpty() && !modelo.getText().isEmpty() && !dominio.getText().isEmpty() &&
                     !anio.getText().isEmpty() && !costoPorKm.getText().isEmpty() && !capacidad.getText().isEmpty()) {
                 GestorCamiones.getGestor().crear(marca.getText(), modelo.getText(), Double.valueOf(dominio.getText()),
-                        Integer.valueOf(anio.getText()), Double.valueOf(costoPorKm.getText()), Double.valueOf(capacidad.getText()), liquido.isSelected());
+                        Integer.valueOf(anio.getText()), Double.valueOf(costoPorKm.getText()), Integer.valueOf(capacidad.getText()), liquido.isSelected());
                 popUpCorrecto();
                 PantallaCamiones.getSingle().actualizarTablaCamiones();
                actualizarValores();
